Start fixing this mess.
- Reworked and simplified installer and config file.
- Fixed startup code.

+++ install.lua
@@ -1,4 +1,52 @@
 
+--[[
+	This program is a generic installer for ComputerCraft. It needs an
+	installer_cfg.lua configuration file to work properly. This file tells the
+	installer from where to download the files and where to store them on the
+	computer.
+
+	Command line options are :
+		uninstall  The default behaviour is install. Uninstall reverses the
+		           actions taken in the install phase.
+		           The deault behaviour is install.
+
+		install    Install all the files specified in the installer_cfg.
+		           Directories will be created as needed
+
+		update     Reinstall only the sections where ["update always"] is False.
+		           Some parts of the package may be stable and not require
+		           frequent updates. To make updates faster this allows us to
+		           ignore them
+
+		update-all Re-Install all the files. This is equivalent to just using
+		           "install".
+
+	installer_cfg.lua documentation
+		a valid installer_cfg.lua file consists of :
+
+			return {  the table with all the configurations  }
+
+		the table should contain a key called ["sections"] that contains a table of sections.
+		each section must have the following keys :
+			["source directory"]       Where to download the files from.
+			["destination directory"]  Where to save the files
+			["update always"]
+			["files"]                  An array of filenames
+
+		to download a file the installer concatenates ["source directory"] and
+		the filenames from ["files"]. There is no way to rename files yet. You
+		can't download a file from pastebin and store it under a  different
+		name.
+
+        ["update always"] is useful if your api contains files that rarely
+        change. For example the terrapin API collection pulls in the Penlight
+        libraries. These libraries are stable now. We don't need to download a
+        new copy each time we run the updater so we set ["update always"] to
+        false. To force the installer to update them we use --all option.
+
+]]
+
+
 local function log(string)
 	local logfile = assert(fs.open("install_log.txt", "a"))
 	logfile.write(string .. "\n")
@@ -7,11 +55,11 @@ end
 
 local function saveFile(path_on_server, path_on_client)
 	local server_file = assert(
-		http.get(path_on_server), 
+		http.get(path_on_server),
 		"failed to download file " .. path_on_server, 2
 	)
 	local client_file = assert(
-		fs.open(path_on_client, "w"), 
+		fs.open(path_on_client, "w"),
 		"failed to open file " .. path_on_client, 2
 	)
 
@@ -24,7 +72,7 @@ end
 local function tbl_as_str(tbl)
 	if tbl then
 		local str = "{\n\t"
-		for k,v in pairs(tbl) do 
+		for k,v in pairs(tbl) do
 			if type(v) == "table" then
 				str = str .. "[" .. k .. "] = " .. tbl_as_str(v)
 			elseif type(v) == "boolean" then
@@ -34,11 +82,11 @@ local function tbl_as_str(tbl)
 					str = str .. (k .. " - false\n\t")
 				end
 			else
-			 	str = str .. (k .. " - " .. v .. "\n\t") 
+			 	str = str .. (k .. " - " .. v .. "\n\t")
 			end
 		end
 		return str .. "}"
-	else 
+	else
 		return "nil"
 	end
 end
@@ -48,13 +96,16 @@ local function uninstallSection(section_name, section)
 	log("Uninstalling " .. section_name)
 	log("section = " .. tbl_as_str(section))
 
+
 	if section["destination directory"] == "/" then
 		for _, file in ipairs(section["files"]) do
 			fs.delete( fs.combine(section["destination directory"], file) )
 		end
-	else
-		fs.delete(section["destination directory"])
+		return
 	end
+
+	-- TODO : Check that the directory is empty before deleting the entire directory
+	fs.delete(section["destination directory"])
 end
 
 local function installSection(section_name, section, base_dir)
@@ -112,73 +163,66 @@ log("Started installer with options : " .. tbl_as_str(options))
 -- clear screen
 term.clear()
 term.setCursorPos(1,1)
-print("tInstaller starting.\n")
+print("Installer starting.\n")
 
 -- check that the configuration file exists. Without this file the installer is useless
 if not( fs.exists("installer_cfg.lua") ) then
 	error("The installer needs an installer_cfg.lua file to continue.")
 end
 
--- load the configuration file 
-local installer_cfg, errors = loadfile("installer_cfg.lua")()
-if not installer_cfg then
+-- load the configuration file
+local installer_cfg
+local cfg_file, errors = loadfile("installer_cfg.lua")
+if errors then
 	error("Could not open installer_cfg.lua. " .. errors)
+else
+	installer_cfg = cfg_file()
 end
 
--- check the mode for the installer. At least one option from 
--- install, update-all or update must be present
-if not(options["install"]) and not(options["update"]) and not(options["update-all"]) then
-	options["install"] = true
-end
+log("Installer startup succesful")
 
-log("Startup succesful")
- 
--- check for traces of previous installations :
--- this block is all that will be run if uninstall mode is set.
-if fs.exists(installer_cfg["base install directory"]) then
-	-- If we are in install mode then delete all the files.
-	-- we will overrite them anyway
-	-- If we are in uninstall mode we want to get rid of them anyway.
-	if options["install"] or options["uninstall"] then
-		-- Warn the use that we are about to uninstall the apis. 
-		if not(options["--force"]) then
-			io.write("A previous installation of terrapin was detected. All " ..
-				"content in the " .. installer_cfg["base install directory"]  ..
-				" will be removed. Continue ? (y/n)"
-			)
+-- If the uninstall option is set then we uninstall everything and return.
 
-			local res = io.read()
-
-			if not (res == "y" or res == "Y") then
-				io.write("\n\n Exiting installer ...\n")
-				return
-			end
-		end
-
-		-- If we are still here the user is Ok with /terrapin disappearing
-		fs.delete(installer_cfg["base install directory"])
-	elseif options["update"] then
-		log("Performing simple update. ")
-		for section_name, section in pairs(installer_cfg["sections"]) do
-			if section["update always"] then
-				uninstallSection(section_name, section)
-			end
-		end
-	else
-		error("No installer mode option present")
+if options["uninstall"] then
+	for section, section_name in pairs(installer_cfg) do
+		uninstallSection(section, section_name)
 	end
+
+	return 0
+end
+
+-- Detect previous installations :
+-- We are either installing or updating. If we are updating then Cool, if we
+-- are installing there might be an error
+if fs.exists(installer_cfg["base install directory"]) then
+	if options["install"] and not(options["--force"]) then
+		io.write("A previous installation of terrapin was detected. All " ..
+			"content in the " .. installer_cfg["base install directory"]  ..
+			" will be removed. Continue ? (y/n)"
+		)
+
+		local res = io.read()
+		if not (res == "y" or res == "Y") then
+			io.write("\n\n Exiting installer ...\n")
+			return
+		end
+	end
+elseif options["update"] or options["update-all"] then
+	io.write(
+		"Unable to locate an existing installation of Terrapin API " ..
+		"Collection. ABORTING."
+	)
 end
 
 log("Previous installation detection succesful. Starting installation")
 
--- Now we can proceed witht the installation. 
+-- Now we can proceed witht the installation.
 fs.makeDir(installer_cfg["base install directory"])
-
 if options["install"] or options["update-all"] then
 	for section_name, section in pairs(installer_cfg["sections"]) do
 		installSection(section_name, section)
 	end
-elseif options["update"] then 
+elseif options["update"] then
 	for section_name, section in pairs(installer_cfg["sections"]) do
 		if section["update always"] then
 			installSection(section_name, section)
@@ -188,11 +232,16 @@ else
 	error("no installer option mode present")
 end
 
+
 log("Install succesful")
 
 if not uninstall_successful then
-	print("The installer detected that some files were added to the terraopin directories. " ..
-		  "These files have been left in place.")
+	print(
+		"The installer detected that some files were added to the terrapin "..
+		"directories. These files have been left in place."
+	)
 end
 
 print "\n\nInstall Succesful"
+
+os.reboot()
